+++ include/sacn/universe_discovery.h
@@ -217,12 +217,12 @@ typedef struct SacnUniverseDiscoveryConfig
   /** The maximum number of sources this listener will record.  It is recommended that applications using dynamic
      memory use #SACN_UNIVERSE_DISCOVERY_INFINITE for this value. This parameter is ignored when configured to use
      static memory -- #SACN_UNIVERSE_DISCOVERY_MAX_SOURCES is used instead.*/
-  size_t source_count_max;
+  int source_count_max;
 
   /** The maximum number of universes this listener will record for a source.  It is recommended that applications using
      dynamic memory use #SACN_UNIVERSE_DISCOVERY_INFINITE for this value. This parameter is ignored when configured to
      use static memory -- #SACN_UNIVERSE_DISCOVERY_MAX_UNIVERSES_PER_SOURCE is used instead.*/
-  size_t universes_per_source_max;
+  int universes_per_source_max;
 } SacnUniverseDiscoveryConfig;
 
 /** A default-value initializer for an SacnUniverseDiscoveryConfig struct. */
